inventory-sync/requirements.txt
mysql-connector-python==8.0.33
redis==5.0.1